+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectOverviewViewModel.groovy
@@ -1,11 +1,15 @@
 package life.qbic.portal.sampletracking.components.projectoverview
 
 import groovy.beans.Bindable
+import java.util.Collections
+
+import life.qbic.business.project.Project
 import life.qbic.business.project.subscribe.Subscriber
-import life.qbic.datamodel.dtos.projectmanagement.Project
 import life.qbic.portal.sampletracking.communication.Topic
 import life.qbic.portal.sampletracking.resource.ResourceService
 import life.qbic.portal.sampletracking.resource.status.StatusCount
+import life.qbic.portal.sampletracking.components.projectoverview.LastChangedComparator
+import life.qbic.portal.sampletracking.components.projectoverview.LastChangedComparator.SortOrder
 
 /**
  * <h1>ViewModel for the {@link ProjectOverviewView}</h1>
@@ -38,9 +42,12 @@ class ProjectOverviewViewModel {
         projectResourceService.iterator().each { Project project ->
             addProject(project)
         }
+        Collections.sort(projectOverviews, new LastChangedComparator(SortOrder.DESCENDING))
+
         statusCountService.iterator().each { StatusCount statusCount ->
             updateSamplesReceived(statusCount)
             updateSamplesFailedQc(statusCount)
+
             updateSamplesLibraryPrepFinished(statusCount)
             updateDataAvailable(statusCount)
         }
@@ -51,6 +58,7 @@ class ProjectOverviewViewModel {
         this.projectResourceService.subscribe({ removeProject(it) }, Topic.PROJECT_REMOVED)
 
         this.statusCountService.subscribe({ updateSamplesReceived(it) }, Topic.SAMPLE_RECEIVED_COUNT_UPDATE)
+        this.statusCountService.subscribe({ updateSamplesPassedQc(it) }, Topic.SAMPLE_PASSED_QC_COUNT_UPDATE)
         this.statusCountService.subscribe({ updateSamplesFailedQc(it) }, Topic.SAMPLE_FAILED_QC_COUNT_UPDATE)
         this.statusCountService.subscribe({ updateDataAvailable(it) }, Topic.SAMPLE_DATA_AVAILABLE_COUNT_UPDATE)
         this.statusCountService.subscribe({ updateSamplesLibraryPrepFinished(it) }, Topic.SAMPLE_LIBRARY_PREP_FINISHED)
@@ -61,38 +69,55 @@ class ProjectOverviewViewModel {
     }
 
     private void removeProject(Project project) {
-        ProjectSummary projectOverview = projectOverviews.collect { it as ProjectSummary }.find { it ->
-            (it as ProjectSummary).code == project.projectId.projectCode.toString()
+        List<ProjectSummary> summaries = projectOverviews as List<ProjectSummary>
+        ProjectSummary projectOverview = summaries.find { it ->
+            it.code == project.code
         }
         projectOverviews.remove(projectOverview)
     }
 
-
     private void updateSamplesReceived(StatusCount statusCount) {
         ProjectSummary summary = getProjectSummary(statusCount.projectCode)
-        summary.samplesReceived = statusCount.count
+        summary.samplesReceived.passingSamples = statusCount.count
+
         int totalSampleCount = statusCount.totalSampleCount
+        summary.samplesReceived.totalSampleCount = totalSampleCount
         summary.totalSampleCount = totalSampleCount
     }
 
     private void updateDataAvailable(StatusCount statusCount) {
         ProjectSummary summary = getProjectSummary(statusCount.projectCode)
-        summary.sampleDataAvailable = statusCount.count
+        summary.sampleDataAvailable.passingSamples = statusCount.count
+
         int totalSampleCount = statusCount.totalSampleCount
+        summary.sampleDataAvailable.totalSampleCount = totalSampleCount
+        summary.totalSampleCount = totalSampleCount
+    }
+
+    private void updateSamplesPassedQc(StatusCount statusCount) {
+        ProjectSummary summary = getProjectSummary(statusCount.projectCode)
+        summary.samplesQc.passingSamples = statusCount.count
+
+        int totalSampleCount = statusCount.totalSampleCount
+        summary.samplesQc.totalSampleCount = totalSampleCount
         summary.totalSampleCount = totalSampleCount
     }
 
     private void updateSamplesFailedQc(StatusCount statusCount) {
         ProjectSummary summary = getProjectSummary(statusCount.projectCode)
-        summary.samplesQcFailed = statusCount.count
+        summary.samplesQc.failingSamples = statusCount.count
+
         int totalSampleCount = statusCount.totalSampleCount
+        summary.samplesQc.totalSampleCount = totalSampleCount
         summary.totalSampleCount = totalSampleCount
     }
 
     private void updateSamplesLibraryPrepFinished(StatusCount statusCount) {
         ProjectSummary summary = getProjectSummary(statusCount.projectCode)
-        summary.samplesLibraryPrepFinished = statusCount.count
+        summary.samplesLibraryPrepFinished.passingSamples = statusCount.count
+
         int totalSampleCount = statusCount.totalSampleCount
+        summary.samplesLibraryPrepFinished.totalSampleCount = totalSampleCount
         summary.totalSampleCount = totalSampleCount
     }