+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/resource/status/StatusCountResourceService.groovy
@@ -15,6 +15,7 @@ class StatusCountResourceService extends ResourceService<StatusCount>{
 
     StatusCountResourceService() {
         this.addTopic(Topic.SAMPLE_RECEIVED_COUNT_UPDATE)
+        this.addTopic(Topic.SAMPLE_PASSED_QC_COUNT_UPDATE)
         this.addTopic(Topic.SAMPLE_FAILED_QC_COUNT_UPDATE)
         this.addTopic(Topic.SAMPLE_DATA_AVAILABLE_COUNT_UPDATE)
         this.addTopic(Topic.SAMPLE_LIBRARY_PREP_FINISHED)
@@ -31,6 +32,10 @@ class StatusCountResourceService extends ResourceService<StatusCount>{
         switch (statusCount.status) {
             case Status.SAMPLE_RECEIVED:
                 publish(statusCount, Topic.SAMPLE_RECEIVED_COUNT_UPDATE)
+                content.add(statusCount)
+                break
+            case Status.SAMPLE_QC_PASS:
+                publish(statusCount, Topic.SAMPLE_PASSED_QC_COUNT_UPDATE)
                 content.add(statusCount)
                 break
             case Status.SAMPLE_QC_FAIL:
@@ -62,6 +67,10 @@ class StatusCountResourceService extends ResourceService<StatusCount>{
         switch (statusCount.status) {
             case Status.SAMPLE_RECEIVED:
                 publish(statusCount, Topic.SAMPLE_RECEIVED_COUNT_UPDATE)
+                content.remove(statusCount)
+                break
+            case Status.SAMPLE_QC_PASS:
+                publish(statusCount, Topic.SAMPLE_PASSED_QC_COUNT_UPDATE)
                 content.remove(statusCount)
                 break
             case Status.SAMPLE_QC_FAIL: