+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/count/CountSamples.groovy
@@ -49,7 +49,7 @@ class CountSamples implements CountSamplesInput{
     } catch (DataSourceException dataSourceException) {
       output.failedExecution(dataSourceException.getMessage())
     } catch (Exception ignored) {
-      output.failedExecution("Could not count received samples.")
+      output.failedExecution("Could not count samples.")
     }
   }